Update paragraph overlap algorithm
When the paragraph styles have same range ends, we need to make sure we clear the stack correctly. It has been done when we cleared the stack at the end of AnnotatedString builder when there was no more paragraph styles to iterate over. But this wasn't handled correctly when we cleared the stack when the new paragraph style arrived.

Test: AnnotatedStringTest
Change-Id: I224984e55eb6fb25736ca0d9ba7054aafaa247c8