Update shape tokens API to be aligned with Compose
Shape object now holds the actual values for Corner radius, instead of
being a map with enums, similarly to ColorScheme.

The difference is that for now, we don't want to expose developers
overriding default values for corners, as it could break the layouts and
components on non-scrolling screen. But this easily allow us to do that
in future, after consideration with UX.

Fixes: 369350414
Test: Updated
Relnote: "Updated Material3 shape to be a class with fields that hold the actual Corner value, same as in Wear Compose."
Change-Id: Ied8cd17340a5d595b8a5ae82540c10760ff2ad95