Fix update_studio script to work with pluginz change
aosp/3305276 missed one pluginz->plugin update for importMaven.

It also created a second line starting with 'androidGradlePlugin = ' in gradle/libs.versions.toml, and the update_studio script started replacing that line in addition to the line that defines the version. This update the regex to only match the version line.

Test: `development/update_studio.sh` runs successfully after updating version numbers
Change-Id: I29d9c4c8de26fd4d6bed789f52674d3cede27878