Trace Deep Links - Insight support
Test: TraceDeepLinkTest
Test: SmallListScrollBenchmark
Bug: 376471150

Now, when TraceDeepLink is used, it's possible to emphasize part of
the trace(selectionParams).

Note that this is currently only supported for insights, though follow
up work should enable this for macrobenchmark metrics as well.

In the case of startup insights, we only embed basic selectionParams
currently. Follow up work in the insights from perfetto should embed
this information, so macrobench can wire it up to deep links more
appropriately.

This change additionally enables startup insights for internal
androidx startup benchmarks.

Change-Id: I294d2a356f48ff9c0f7c901203261e08ddcc68e2