Clear initialization Future on ProcessCameraProvider failure
This change ensures that the ListenableFuture used for
ProcessCameraProvider initialization is cleared if initialization
fails. This allows for proper re-initialization the next
time ProcessCameraProvider.getInstance() is called.

Previously, if ProcessCameraProvider initialization failed,
the ListenableFuture associated with it was not cleared.
This could lead to subsequent calls to
ProcessCameraProvider.getInstance() returning the same failed
ListenableFuture, preventing successful re-initialization.

This CL modifies the code to call shutdown internally whenever
ProcessCameraProvider initialization fails. This ensures that
future calls to ProcessCameraProvider.getInstance() will trigger
a fresh initialization attempt.

Bug: 373284521
Test: Device tests in CameraX lab
Change-Id: I37c8430cc7380a9e5b29601aa60562351d0a55e0