Expose LifecycleCameraProvider as public APIs.
Previously ProcessCameraProvider is a singelton, which only allows being
configured once for the process. This CL adds APIs to allow camera
provider to be instantiated so that it can be configured differently
for different usages. For example, if an application wants to open the
camera on a virtual device, the camera provider has to be reconfigured
with a different context.

RelNote: "Added LifecycleCameraProvider, which is a camera provider that
can be instantiated with different configurations for features such as
accessing the camera of a virtual device by configuring with a
customized context."

Bug: b/343347002
Test: ./gradlew camera:camera-lifecycle:updateApi,
./gradlew camera:camera-lifecycle:build

Change-Id: Ia29928816ac5750f2875473fe600b62d99185047