Introduce ViewGroupCompat#installCompatInsetsDispatch
The API makes the given view dispatch insets to its descendants in a way
compatible with API 30+ (Android R+) during the insets dispatching.

Before API 30, if a view consumes insets during the insets dispatching,
the whole dispatching process will stop. So not only the child views of
the consuming view won’t receive the insets, but neither will the rest
of its sibling views and their child views.

Since API 30, the problem has been fixed that only the child views of
the consuming view won’t receive the insets, and the sibling views won’t
get affected.

The API aims to fix the issue for apps running on a platform before API
30.

Bug: 369883070
Relnote: "Add a new API `ViewGroupCompat#installCompatInsetsDispatch`
for developers to fix the insets dispatching issue before API 30"
Test: ViewGroupCompatTest

Change-Id: I111590881991b94e5cd5f2100bf88dd4468bc781