Add `showAutofillDialog` to Autofill implementation
This CL adds in a `showAutofillDialog` call whenever a component is entered. If conditions for autofill suggestions in a dialog-style UI are met (as determined by the autofill framework), then the pop-up dialog is displayed. Since the autofill framework has heuristics to know if the conditions are met, Compose will always call `showAutofillDialog`, and the UI will be triggered when needed.

Bug: 349386998
Test: autofillManager_showAutofillDialog_previousFocusFalse and others
Change-Id: I0a67ef4c2badcabcfff9fa695b4c0ceb9a9d59c1