Raise CameraDevice.close() timeout and log the error instead
Previously Camera2DeviceCloser should've thrown
TimeoutCancellationException from withTimeout. However, it became clear
that wasn't the case, and aosp/3288148 was created to mitigate the
error.

Based on production crash date, it seems that even with a 5s timeout, it
is still not an uncommon occurrence that camera device closure times
out. Given the camera framework also has a 5s timeout for camera device
closure, or more specifically for camera device dropping inflight
requests, raise the timeout to 8s.

Now given the timeout at 8s, if it still times out, log the error that
clearly gives an indication that the cameras may not work normally
again.

Bug: 374115214
Test: CameraPipe test app
Change-Id: If563285fb1ba9374d33afaae39684face988bfde