Add CompositionRegistrationObserver API
This CL introduces a new tracing API requested by TikTok (but not
/that/ TikTok).

The introduced API allows for observing a Recomposer to receive a
callback when a new composition is created within it. This makes it
easier to automatically install tooling into subcompositions or
in interwoven interop use cases, for example.

Fixes: b/366038833
Test: RecomposerObserverTest
Relnote: """
    Adds tooling APIs to observe the creation of new compositions within
    a Recomposer
    """
Change-Id: Iaeb7e218aa949ebf1c061b311a3c4383cf5c2728