Fix failing Modifier.focusable() test
After aosp/3296711 landed, it broke the following test when the feature flag was turned on
androidx.tv.material3.NavigationDrawerTest.navigationDrawer_parentContainerGainsFocus_onBackPress()

This failure was caused by the eager calculation of semantics configuration that was
introduced in aosp/3296711. It surfaced an existing bug in Modifier.focusable, where
invalidateSemantics() was called before the focus state was updated. This didn’t matter
when we were calculating semantics configuration lazily, but is an issue now.

This issue only affects an older version of Modifier.focusable that uses a now deprecated
FocusTargetModifierNode function with no parameters. The newer implementation of
Modifier.focusable does not have this issue because it does not keep a local copy of focus
state, but instead reads focus state from FocusTargetNode directly.

Test: Turned the flag on for the failing test and verified that presubmits pass.
Then turned the flag off again before sending out the CL for review.
Bug: 272068594
Change-Id: Icbc366eb6bedcd9d8843066967fe5365e5170c08