Add support for expandable component state control
- Introduced a floatingToolbarVerticalNestedScroll Modifier that can
be attached to a scrollable container and update the expansion state
of a floating toolbar when the scroll is crossing a threshold.
- Applied this Modifier to trigger an expansion or a collapse of a
floating toolbar in the samples.

Bug: 373460660
Test: Added ScrollAwareToggleTest
Relnote: "Added a new floatingToolbarVerticalNestedScroll that can be
attached to a scrollable container to update the floating toolbar
expansion state based on a scroll motion that crosses a threshold."

Change-Id: I6d65f467fb88e31233dbb9a266d8aed70d5aea32