Fix nodes fixup in ordered collections
Adding numerous items creating key collisions could lead to a certain
traversal sequence in dropDeletes() that could in turn lead to the
corruption of the internal linked list. That corruption could later
cause the wrong eviction candidate to be used when trimming the size
of the cache, leading to a null pointer exception when retrieving
the eviction candidate key.

To create that situation, dropDeletes() would first need to swap two
items in the data structure, then re-process one of the items and
re-assign it to its current location. This would cause the loop to
erase the relocation mapping, creating the aforementioned corruption.

This case is now covered by two test cases: the existing test called
hashCollisions() (now renamed sequentialHashCollisions()), updated
with a larger iteration count to create the problem, and a new test
called randomizedHashCollisions() which does the same, but using non
sequential collision-provoking keys.

Fixes: 375607736
Relnote: Fixed a crash that could occur in SieveCache and OrderedScatterSet when adding numerous items causing repeated key collisions
Test: OrderedScatterSetTest, SieveCacheTest
Change-Id: Iaaf3d36d6adb491daf564d5625810446b6f5d321