Remove XFunSpec.Builder.name and XTypeSpec.className
XFunSpec.Builder has access to the underlying JavaPoet or KotlinPoet
type so there is no need to pass in a separate name since it can be
accessed from the actual types.

Similarly, XTypeSpec has access to the name in the underlying type.
In addition, XTypeSpec should be using just name (String) rather than
XClassName since the actual class name of an XTypeSpec depends on where
it is added (e.g. if it's nested in another type or not) so we should
not try to pass in the XClassName ahead of time.

Test: Relies on pre-existing test
Change-Id: I1d0bb8c29755d0e127f0f7f1fffb2f02ec19b0df