Change PausableComposition API from feedback from API council
- The `flag` paramter of `shouldExecute()` was documentd.
- `PausableComposition` and `PausedComposition` changed to be
sealed interfaces.
- The parameter `PausableComposition.resume()` was changed to
be a `fun interface` instead of a normal lambda to avoid
boxing the `Boolean` result allowing slightly better code
generation.

Test: existing tests, ./gradlew :compose:r:r:test
Fixes: 376516879
Relnote: """Changes to the PausableComposition API.

These breaking changes with respect to previous releases of
1.8.0-alpha (since 1.8.0-alpha02) but, as these are new APIs
in 1.8.0, backwards compatibility is not maintained as the APIs
are still in alpha."""

Change-Id: I39819fd1a6509973a0008cd4fdec8a892829a301