Refactors pagination logic out of PdfView
Logic is moved into an event-driven, state-producing PaginationManager. PdfView now interacts exclusively with PaginationManager and not PaginationModel.

PaginationModel was updated so that the reach property is properly 0-indexed, and its tests were updated accordingly.

This fixes the exposed implementation detail wherein PdfView.visiblePagesCount would be 2 with 0 loaded pages.

Use Espresso and ActivityScenario for PdfViewPaginationTest and PdfViewRenderingTest. Previously these tests manually laid out and measured a PdfView without attaching it to any Window.

Implement scrollToPage and scrollToPosition public APIs. This will likely be extracted to a separate change, but I was using the feature to sort of stress test this refactor.

Test: ./gradlew :pdf:pdf-viewer:(test|cAT), manual
Change-Id: Idb166b4e075ed893243f21fb7e882321164e02cd