Stop crash when moving focus to ComposeView with no focusable content
Fixes: 369256395
Fixes: 378570682

Compose advertises itself as focusable. When another View finds the
ComposeView and requests focus on it, it is expected to accept the
focus and can crash if it doesn't. This CL returns true when that
happens and tries to move focus to the next item if possible.

When moving focus between ComposeViews using "Tab," the
focus was not changing. This CL fixes that problem.

Test: new test, manual testing
Change-Id: I372b6974e1b4714f825c208e683aea8d50e19116