Fix back stack changes during host lifecycle changes
When we dispatch lifecycle changes to destinations
because of back stack changes, we operate on a copy
of the back stack to avoid reentrant changes -
lifecycle observers that call navigate / popBackStack.

Now, we do the same for host lifecycle changes,
ensuring that any lifecycle observers that are triggered
when the host lifecycle changes state also avoid
any reentrant issues due to changing the back stack
during the dispatch of the lifecycle change.

Relnote: "Fixed a `ConcurrentModificationException` that
could occur when a `LifecycleObserver` attached to a
`NavBackStackEntry` triggers a change to the back stack
when the host `LifecycleOwner` such as the containing
Activity or Fragment changes its lifecycle state."
Test: new NavControllerTest
BUG: 377537292

Change-Id: Ia9494e9967aa45a7613c49d7530f524f5400d1ca