Only unbind the cameras that were created by the same camera provider.
Previously, when unbind or shutdown is called, it unbinds all the
LifecycleCameras because they are expected to be created by the same
camera provider, which is a singleton. But we are going to allow
creating multiple camera providers. So each provider should only unbind
the cameras that were originally bound by itself.

Bug: b/343347002
Test: ./gradlew camera:camera-lifecycle:build
Change-Id: Id3330784a4398dee65ae923ad091b7cd57f0dc5a