Add base classes for Nav3
Adding the building blocks for the nav3 library. The Record class
provides a connection between a key and some composable content. The
NavContentWrapper allows passing of information to either the back stack
or a record. The NavWrapperManager handles the interactions between the
list of NavContentWrappers and then currently active content.

RelNote: N/A
Test: Added RecordTest
Bug: 374890821
Change-Id: I58961b9d9e1dd5c9cf81f2dc0f744810ff51380c