Remove gesture drawing logic from TestView
Gesture drawing was added to ensure there are no differences between
touch transfer in Z-above and Z-below for SandboxedSdkViews. As Z-below
doesn't support this at the moment, this logic is removed from the test app for now.

Additionally, removing the logic solves the problem of slow clicks on
non-webview banner ads. This was caused by calling 'invalidate()' on
each touch event.

Bug: 367051054
Test: manual
Change-Id: I43e22d8341615589c9f061b408515647e852cddc