Fix ScrollIndicator position in LazyColumn and ScalingLazyColumn with ContentPadding
This CL addresses an issue where using contentPadding in LazyColumn or ScalingLazyColumn did not affect the position of the ScrollIndicator. This resulted in unexpected behavior where the list could be scrolled, but the ScrollIndicator did not update.

This fix specifically addresses contentPadding and does not resolve issues related to AutoCentering padding (b/378470782).

Bug: 368270238
Test: ScrollIndicatorTest.kt
Relnote: "Fix ScrollIndicator position in LazyColumn and ScalingLazyColumn with ContentPadding."
Change-Id: I2bc51a0d5b12916e3d560bf9b2e9119601fb21a9