Use reified serializer for EntityBundle and FtsEntityBundle for xref to work
This overload works with full type information, including type arguments and nullability, and is a recommended way to retrieve a serializer.

Motivation:
I needed to change SchemaBundle.kt as xref generation fails
to interpret that `.serializer()` is a generated function from the
serialization plugin. The pipeline does not have -Xplugin argument
passed in (It doesn't allows it).

To fix, we trick the code to use reified serializer

Change-Id: If70ccb03d150b743b4bb8d39d6a47cf092b77cc9
BUG: 374948517