Update package names in compose lint tests
These tests all had packages named `foo` and also a top level property `foo`. In a new android lint version, the FULLY_QUALIFIED test mode started fully qualifying references to the property `foo` as `foo.foo`, which caused issues because the first `foo` was getting parsed as the property instead of the package. This renames the package to avoid the ambiguity.

Test: ./gradlew test
Bug: 378110980
Change-Id: Iee0ec1b802ffe84b284abf4fa8f880dd131d6ce1