Remove duplicate definition in lifecycle lint test stubs
The `REPEAT_ON_LIFECYCLE_STUB` contained a definition of `androidx.lifecycle.Lifecycle`, but it was only used along with the `LIFECYCLE_STUB`, which already contains a definition of the class. These duplicate definitions were previously not an issue, but a new android lint version started failing on tests using both of these stubs.

Bug: 378110980
Test: RepeatOnLifecycleDetectorTest
Change-Id: I15424f090e947a605ccc1909f2466802cf43a43f