Changes to the removeChild API.
Change removeChild API for AndNode and OrNode to take Node instead of an
index. This avoids clients from having to maintain indices whenever the
list changes and provides an better way of removing from a list by using
a node instead of having to find the index first and then passing the
index.

Added an API that returns the index of a given Node in a list.

Bug: 356661489
Bug: 356662352
Bug: 369826552

Relnote: Change to removeChild API signature from int to Node.
Test: "Change removeChild tests to check for nullness, node not existing.
Add test to make sure removeChild removes the node."

Change-Id: I47aa5b7b396d425d7e35b066da254468676e37d8