Fix bug on scrolTransform when adding/removing/movign items.
When used on a TLC and adding removing items, the 'remember's inside
Modifier.scrollTransform could be reused based on index in the list,
which breaks when items are modified. Adding `scope` to the keys
used on the remembers fixes the issue.
Relnote: Fix bug on scrolTransform when adding/removing/movign items.
Test: Manual, added sample to exercise adding/removals.
Bug: 377896252

Change-Id: I6830fe9d9df3645a822eb3e1b352aad1a1def2d6