Update AndroidX @Discouraged retention to BINARY
Usage of androidx.annotation.Discouraged does not raise lint warnings as
expected because the annotation is not retained in binary output. This
change updates the annotation's retention policy from SOURCE (discarded
by the compiler) to BINARY (stored in binary output; not visible for
reflection).

This change corresponds to the stub change at aosp/3361226.

Test: Presubmit
Bug: 375937557
Relnote: Updated @Discouraged retention policy.
Change-Id: Id2f85e1b47c9f8555352b5a29b1df70d6067af20