Improve RemoteActivityHelper documentation.
Since RemoteActivityHelper is limited in which intents it can deliver,
and it is sometimes unclear what the requirements are and what happens
to additional options that might be set on the intent.

This commit attempts to make it clearer in the documentation that
additional things, like `extras`, are not copied over. This has come
up occasionally as something people had questions about.

Bug: 379724614
Test: comments only change
Change-Id: Ie65cb19993129e93cc2dcc1bbfad963c9c776d51