Properly clear SaveableStateNavContentWrapper state
When the back stack is changed, we should immediately clear the state in
SaveableStatenavContentWrapper. Currently with a `LaunchedEffect` the
state is clear after the new incoming back stack has already come in and
been added to the refCount, which causes us to immediately clear our new
state.

Swapping over to a `DisposableEffect` and moving it above the one based
on the key ensures that we do backstack operations (i.e. clearing) as
soon as the back stack changes.

Test: NavDisplayTest
Bug: 379898461
Change-Id: I9868ac4d2157e488172acd87d514713c0691dfcd