Implement compose stack traces for applyChanges()
Handles crashes in the apply phase by connecting crash execution place to a location in composition.

For changelist:
Updates Operation interface to have an overload that accept stack trace context. The context connects operation invocation place (usually marked by the slot writer position) with the parent context to provide a full trace to the root.
For most operations it is enough to just use current location of the writer, which is handled by the default implementation. Some operations (e.g. node insertion fixups) require additional context stitching, so they are implemented separately.

For remember observer events:
Looks up the location of the node / observer holder in the slot table. This implementation does not provide context for `onForgotten` callbacks (or `onDetach`/`onRelease` of the node) at the moment.

Test: ErrorTraceTests
Bug: 354163858
Change-Id: I6ccb66004099d895ddffecacb85f582c9f7c9cc8