Launch MagnifierNode coroutine undispatched
When MagnifierNode is attached it launches a coroutine. In a test this
happens unconfined and thus passes the assertion that it catches the
first draw signal. However, in production the launch is dispatched on
the main dispatcher, and it could run after draw() has been called
[citation needed].

Bug: 380878212
Test: "Can't reproduce without support for StandardTestDispatcher.
Verified by looking at logs while running MagnifierDemos."

Change-Id: I6c121c058fde40b1e5675195b7ed2417ef74bd63