Add settings to enable/disable the wrapper route provider
We added a preference fragment to host switch buttons to enable/disable
the simple route provider and the wrapper route provider. In this case,
the settings could be stored in SharedPreference. Also, we can read the
settings from SharedPreference when the app starts and then configure
the demo app accordingly.

Bug: 348357495
Test: Manually tested with the demo app
Change-Id: I50ebc827d019300c94f6ff03b57fd36a7a93f3ec