Fix the app-bars behaviors to detect end of scroll
Apply a fix to all the behaviors to detect the end of scroll in the post
fling function. This will better ensure an app-bar color change when the
content is scrolled all the way.

Bug: 293665988
Test: Manual
Relnote: "Fix the top and bottom app-bar behaviors to reliably change
color when content is scrolled all the way."

Change-Id: Idc4e834695cbd7cb8099f7b63cf21d5b764f1c81