Remove clipping bounds logic from SandboxedSdkView
The clipping bounds logic was used to clip the SandboxedSdkView to
its nearable scrollable parent (or the root content view, if the
SandboxedSdkView does not have a scrollable parent).

This logic is being removed as it does not cover all scrollable
container cases (e.g. HorizontalScrollView), and it has introduced
other issues related to flakiness with
orderProviderUiAboveClientUi(false).

Relnote: """Removed clipping bounds logic from SandboxedSdkView.
    This will cause UX issues if SandboxedSdkView is placed in
    a scrollable container and orderProviderUiAboveClientUi(false) has
    not been called (until the underlying framework bug is fixed)."""
Test: SandboxedSdkViewTest, manual
Bug: 380873060
Change-Id: Id420d880865a6debe6d595dfc2e6e9cd3316b236