Restructure updateAbi task to include kotlin compiler in classpath
aosp/3368960 made the Kotlin compiler a `compileOnly` dependency in `buildSrc` to avoid a warning that comes up when both KGP and the compiler are in the build classpath. Most of the BCV tasks are defined externally in `kotlinx.validation` and allow setting a `runtimeClasspath` which is used to execute the tasks, and aosp/3368960 set the classpath for those tasks which allowed them to keep using kotlin compiler classes.

However, `updateAbi` is defined in `buildSrc` and did not have a way to set the classpath, so it was failing because kotlin compiler classes were not available at runtime. I updated the task to have the same structure as the `kotlinx.validation` tasks so the kotlin compiler will be in the classpath as needed.

Bug: 380892014
Test: `./gradlew updateAbi` passes locally. Also tested throwing a `ParseException` from the BCV code to verify the task output would look the same as before when the try/catch was directly in the UpdateAbiTask.
Change-Id: I0f30b54c1130fcdd6754efe70d84d943dde62202