Move PerfettoTraceProcessor -> TraceProcessor in benchmark-traceprocessor
Bug: 381134564
Test: Tests in benchmark-macro
Test: TraceProcessorBenchmark

Relnote: """Move PerfettoTraceProcessor to TraceProcessor in a new
androidx.benchmark:benchmark-traceprocessor artifact, and make most of
its API non-experimental. Any custom TraceMetric or anything reading
from traces will need to update to the new TraceProcessor import. The
new TraceProcessor API works exactly like the old one, but is a
standalone interface library (somewhat analogous to the androidx.sql
layer from Room) with an Android-specific implementation built into
macrobenchmark. The new artifact can be used on JVM as well, but
currently you'll need to start your own copy of the TraceProcessor
binary and offer a port to connect to it on."""

This is a first step toward writing Metric / Insights code that runs
(and can be tested) on host

Change-Id: I6256342f25d3f00ef864cf14b75a7159c6a40573