Remove unnecessary check in SearchResultsImpl.
This check throws if the caller sets a JoinSpec before the API is
available. This is unnecessary because this check is already enforced in
the SearchSpecToPlatformConverter#toPlatformSearchSpec.

This check was added first historically and was around the time we were
first starting this pattern. So I believe, that it was added and then
never properly cleaned up when the correct check was added to the
converter afterwards.

Test: presubmits
Change-Id: I0d0f6e1d544f944481833108e491b083efaa02cd