Make CameraGraph.close non-blocking and move the timeouts
After aosp/3369087, we should be able to revert the await on shutdown in
Camera2CameraController.detachSessionAndCamera(), called during stop()
and close().

This CL removes the await and moves the timeout handling to the specific
areas where things can block indefinitely, namely abortCaptures()
(b/379855962) and close() (b/307594946).

Here we also adjust the camera device close timeout and handle it
somewhat more nicely, as lab testing showed cases of close timing out.
Based on code traces, it's also determined CameraDevice.close() is the
one that can block, beyond that, onClosed() doesn't get much blocker
ahead of it. Hence we adjust the timeouts to 7s and 2s respectively.

Bug: 379855962
Bug: 307594946
Test: CameraX lab testing.
Change-Id: I7b306b3f0551810c388d159ac1d1627ed6723af3