Disable RestrictTo lint for WebView testapp
The Webkit library has a test app that lives in a separate project.
None of the RestrictTo.Scope values properly allows the test app access
to APIs during development.

This CL disables the lint check for the test app to allow developers to
work on the test app without problems from the lint check.

Fixes: 380293915
Test: Manual testing in downstream test CL
Change-Id: Idb44a6f17861032df7840f2393997e5938f85bcc