Added group by period aggregation fallback.
Aggregator is the main interface that accepts record types and returns
an aggregation result. Could be a single result or a bucketed result.

ResultAggregator is the implementation of Aggregator for non-bucketed
result.

ResultGroupByPeriodAggregator is the implementation of Aggregator for
group by period aggregation result.

AggregationProcessor is the interface for the actual calculation of
aggregation. Each record type overrides this for their aggregation
metrics and then this is passed as an argument to either
ResultAggregator or ResultGroupByPeriodAggregator.

The logic for aggregating each type was added in previous CLs and
refactored into implementations of AggregationProcessor.

Test: Added unit and emulator tests. Verified locally with those tests
as well.
Bug: 326414908

Change-Id: Ic1d372a65bd533c2ca2be5d05451dc86e065d2b6