Address UX feedback for M3 AlertDialog
Due to the way SLC works, there is some inherent bottom spacing,
so unable to specify the exact bottom spacing for confirm and
dismiss buttons as in spec (14%), instead account for the existing
spacing and add a smaller spacing so the dialog looks like the
spec.

Bug: 381283085
Test: N/A
Relnote: "Changes to the UX of AlertDialog, on large screens the
confirm and dismiss buttons are now smaller. There is also increased
spacing underneath the confirm and dismiss buttons."

Change-Id: I4f06620d79f6265e012791fb080d65d7cd9e4a89