Stabilize `invisibleToUser`
A 1.7 cherry-pick (aosp/3341487) to fix b/376479686 calls `invisibleToUser` from the foundation module, but it cannot depend on `invisibleToUser` if it is still experimental, so experimental was removed from `invisibleToUser` in 1.7 (via aosp/3387946). This change removes experimental from `invisibleToUser` in 1.8+ so that it matches the change in 1.7 for binary compatibility.

Leaving relnote blank since `invisibleToUser` is still deprecated in 1.8+

Relnote: N/A

Bug: b/376479686
Test: N/A
Change-Id: I40cdd6a3d335011e890761560a5a3416aa579991