Use Java 8 APIs in buildSrc for project to be indexed.
I tried to add a lint check to make sure we only use Java 8 APIs in buildSrc, but the only way to do so (or I can find out) is to list every single Java 9+ API call and run a FQN search on each PsiElement. I tried to use the "@since" in the docstring as an alterantive but element.docComment always returned null.

I also tried to compile ct.jar and add it to the classpath for indexing but it requires a lot of work. Google3 does it but it involves a few steps involving a couple of JARs that also need to be created to generate ct.jar.

For now, we will know if we are using Java 9+ API if buildSrc project stops being indexed

TESTED: Ran ./gradlew :annotation:annotation:publish and extracted the LICENSE.txt from one of the libs and it had the correct text
Ran ./studiow all from terminal with another studio instance already open and it was able to read the PID and throw the error

Change-Id: I4292a6f34ed6cd7b9f7856165e4244afcc2470b8