Switch JSpecify lint from gradle property to AndroidX extension property
Previously the JSpecify lint check would only run when the `androidx.useJSpecifyAnnotations` property was set in the gradle command. This switches the mechanism for turning on the lint check to a per-project extension property, so it can always be run for projects that have been migrated to JSpecify.

Currently the lint check is always off. It will be switched to default on in the following CL, opting out projects that haven't been migrated yet (before that CL is submitted, some projects that are only present in platform-dev need to be opted out).

Test: ./gradlew lint
Bug: 326456246
Change-Id: I375ef175eb1b576dd4853f9abcae251c0fd76d5d