Finalize the camera device after quirk handling and closing
In an earlier CL, aosp/3371201, a quirk handling mechanism was
introduced where we reopen the camera device, close the capture session
before finally closing the camera device.

In the same CL, we finalize the camera device on the initial close,
but that caused the Surfaces to be invalidated quicker, which would
result in new CameraGraphs getting created quicker as well. This then
has a slight timing side effect where CameraPipe may reopen cameras a
bit more frequently than these legacy devices anticipate, which may
result in a deadlock in the camera framework.

Here this CL moves the finalization till after the quirk handling
procedure.

Bug: 379347826
Test: RecorderTest and OpenCloseCaptureSessionStressTest
Change-Id: Ia1cc00d1e8ba7be03eefb84c97e07ad0c6ff36c9