Resume updateVideoUsage tests
A persistent recording should be stopped explicitly, however currently
there's a race condition when it's stopped immediately after rebinding,
so add a wait for the status event before stopping to ensure the
configuration has completed.

Fix whenUseCaseBoundToNewCameraForPersistentRecording that the use cases
are not correctly bound.

Remove whenLifecycleStoppedBeforeCompletingPersistentRecording since a
persistent recording is not supposed to be stopped when lifecycle stops.

Bug: b/353113961
Test: ./gradlew camera:camera-video:build
Change-Id: Ie361d86c50773ede6464d332761ca58d878f2f39