Support launching UI library test app using extras
Adds a FragmentOptions construct which can be used to launch the manual
test app in a specific configuration for testing. This currently
supports all of the test app options apart from the full screen
activity. For all configuration options, there is a default option in
the case that the option is not specified.

Example usage: adb shell am start -n
androidx.privacysandbox.ui.integration.testapp/.MainActivity
-e "fragment" "pooling-container"
-e "ui-framework" "compose"
-e "z-order" "above"
-e "mediation" "in-app"
--ez "draw-viewability" true

Test: Manual
Bug: 381237843
Change-Id: I172d1b4704140e5b43246db825dd93aa8436b9e4